:hammer: Refactor: 반려동물 상세 페이지를 feature 단위로 분리
- 반려동물 상세 페이지에서 화면 로직을 features/pet-detail로 분리
- 상세 본문, 로딩, 에러 UI를 각각 독립 컴포넌트로 분리
- 종, 성별, 날짜 표시 포맷 로직을 lib 계층으로 이동
- pages/my/PetDetailPage는 라우팅과 레이아웃 연결만 담당하도록 정리

`#29`
