+++ src/features/pet-detail/index.ts
@@ -1,0 +1,3 @@
+export { PetDetailContent } from './ui/PetDetailContent';
+export { PetDetailError } from './ui/PetDetailError';
+export { PetDetailSkeleton } from './ui/PetDetailSkeleton';
